+++ CHANGELOG.md
@@ -1,10 +1,15 @@
 # Changelog
 
+## [1.1.0]
+- Feature: video upload
+- Improvement: show success message when adding video
+
 ## [1.0.1]
 - Bugfix: fixed link to video
 - Bugfix: fixed redirect to 'edit' after 'create'
 - Bugfix: fixed permission check
 - Improvement: user textfilter (full metadata search) instead of title
+- Changes to work with latest (refactored) Opencast plugin version
 
 ## [1.0.0]
 - Initial Release